# internal/booking/api/generated/go/model_booked_room.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Hotel Booking API
 *
 * API for fetching and booking rooms by hotel ID.
 *
 * API version: 1.0.0
 */

package openapi


import (
	"time"
)



type BookedRoom struct {

	ID string `json:"ID,omitempty"`

	HotelID string `json:"HotelID,omitempty"`

	Name string `json:"Name,omitempty"`

	Description string `json:"Description,omitempty"`

	Price int32 `json:"Price,omitempty"`

	Entry time.Time `json:"Entry,omitempty"`

	Exit time.Time `json:"Exit,omitempty"`

	// Email of the guest
	Email string `json:"Email,omitempty"`
}

// AssertBookedRoomRequired checks if the required fields are not zero-ed
func AssertBookedRoomRequired(obj BookedRoom) error {
	return nil
}

// AssertBookedRoomConstraints checks if the values respects the defined constraints
func AssertBookedRoomConstraints(obj BookedRoom) error {
	return nil
}